feat(tests): add new tests for archetype validation and combat mechanics
- Implement tests for tie-breaking logic in archetype ranking.
- Add assertions to validate that weight scaling does not affect combat outcomes.
- Introduce tests for Kendall's tau-b rank agreement in behavioral metrics.
- Enhance combat tests to verify average cooldown and stun application.
- Create a smoke test for multi-run artifact routing to ensure proper file handling.
- Update provenance tests to ensure measurement code is included in artifact stamps.
- Refactor elite selection logic to use a rate-based approach instead of absolute counts.
- Adjust visualization code to ensure stun values are correctly represented as integers.

+++ src/visualization/viewer.py
@@ -140,7 +140,7 @@ def _extract_events(trace: CombatTrace, chars: Tuple[Character, Character]) -> L
                 damage=dmg,
                 hp_before=hp_before / hp_max[defender],
                 hp_after=hp_after / hp_max[defender],
-                stun=float(trace.stun_applied[t, att]),
+                stun=int(trace.stun_applied[t, att]),
                 ko=hp_after <= 0.0,
             ))
     return events
